Part 6-Pytorch with Huggingface/requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/dev.in
#
--find-links https://download.pytorch.org/whl/torch_stable.html

absl-py==1.2.0
    # via tensorboard
addict==2.4.0
    # via -r requirements/dev.in
aiohttp==3.8.1
    # via fsspec
aiosignal==1.2.0
    # via aiohttp
albumentations==1.2.1
    # via -r requirements/dev.in
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
black==22.6.0
    # via -r requirements/dev.in
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   black
    #   safety
    #   wandb
coloredlogs==15.0.1
    # via onnxruntime
cycler==0.11.0
    # via matplotlib
docker-pycreds==0.4.0
    # via wandb
dparse==0.5.1
    # via safety
efficientnet-pytorch==0.7.1
    # via segmentation-models-pytorch
flatbuffers==2.0
    # via onnxruntime
fonttools==4.34.4
    # via matplotlib
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.7.1
    # via pytorch-lightning
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via wandb
google-auth==2.9.1
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
grpcio==1.48.0
    # via tensorboard
humanfriendly==10.0
    # via coloredlogs
hydra-core==1.2.0
    # via -r requirements/dev.in
idna==3.3
    # via
    #   requests
    #   yarl
imageio==2.20.0
    # via scikit-image
importlib-metadata==4.12.0
    # via markdown
importlib-resources==5.9.0
    # via hydra-core
joblib==1.1.0
    # via scikit-learn
kiwisolver==1.4.4
    # via matplotlib
kornia==0.6.6
    # via -r requirements/dev.in
markdown==3.4.1
    # via tensorboard
markupsafe==2.1.1
    # via werkzeug
matplotlib==3.5.2
    # via -r requirements/dev.in
mpmath==1.2.1
    # via sympy
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
munch==2.5.0
    # via pretrainedmodels
mypy-extensions==0.4.3
    # via black
networkx==2.8.5
    # via scikit-image
numpy==1.23.1
    # via
    #   albumentations
    #   imageio
    #   matplotlib
    #   onnx
    #   onnxruntime
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   pytorch-lightning
    #   pywavelets
    #   qudida
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tifffile
    #   torchmetrics
    #   torchvision
oauthlib==3.2.0
    # via requests-oauthlib
omegaconf==2.2.2
    # via hydra-core
onnx==1.12.0
    # via -r requirements/dev.in
onnxruntime==1.12.0
    # via -r requirements/dev.in
opencv-python==4.6.0.66
    # via -r requirements/dev.in
opencv-python-headless==4.6.0.66
    # via
    #   albumentations
    #   qudida
packaging==21.3
    # via
    #   dparse
    #   hydra-core
    #   kornia
    #   matplotlib
    #   onnxruntime
    #   pytorch-lightning
    #   safety
    #   scikit-image
    #   torchmetrics
pandas==1.4.3
    # via -r requirements/dev.in
pathspec==0.9.0
    # via black
pathtools==0.1.2
    # via wandb
pillow==9.2.0
    # via
    #   -r requirements/dev.in
    #   imageio
    #   matplotlib
    #   scikit-image
    #   segmentation-models-pytorch
    #   torchvision
platformdirs==2.5.2
    # via black
pretrainedmodels==0.7.4
    # via segmentation-models-pytorch
promise==2.3
    # via wandb
protobuf==3.19.4
    # via
    #   onnx
    #   onnxruntime
    #   pytorch-lightning
    #   tensorboard
    #   wandb
psutil==5.9.1
    # via wandb
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydeprecate==0.3.2
    # via pytorch-lightning
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytorch-lightning==1.6.5
    # via -r requirements/dev.in
pytz==2022.1
    # via pandas
pywavelets==1.3.0
    # via scikit-image
pyyaml==6.0
    # via
    #   -r requirements/dev.in
    #   albumentations
    #   dparse
    #   omegaconf
    #   pytorch-lightning
    #   wandb
qudida==0.0.4
    # via albumentations
requests==2.28.1
    # via
    #   fsspec
    #   requests-oauthlib
    #   safety
    #   tensorboard
    #   torchvision
    #   wandb
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
ruamel-yaml==0.17.21
    # via safety
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
safety==2.1.1
    # via -r requirements/dev.in
scikit-image==0.19.3
    # via albumentations
scikit-learn==1.1.1
    # via
    #   -r requirements/dev.in
    #   qudida
scipy==1.9.0
    # via
    #   -r requirements/dev.in
    #   albumentations
    #   scikit-image
    #   scikit-learn
segmentation-models-pytorch==0.3.0
    # via -r requirements/dev.in
sentry-sdk==1.9.0
    # via wandb
setproctitle==1.3.0
    # via wandb
shortuuid==1.0.9
    # via wandb
six==1.16.0
    # via
    #   docker-pycreds
    #   google-auth
    #   grpcio
    #   munch
    #   promise
    #   python-dateutil
    #   wandb
smmap==5.0.0
    # via gitdb
sympy==1.10.1
    # via onnxruntime
tensorboard==2.9.1
    # via pytorch-lightning
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
threadpoolctl==3.1.0
    # via scikit-learn
tifffile==2022.7.28
    # via scikit-image
timm==0.4.12
    # via
    #   -r requirements/dev.in
    #   segmentation-models-pytorch
toml==0.10.2
    # via dparse
tomli==2.0.1
    # via black
torch==1.11.0+cu113
    # via
    #   -r requirements/dev.in
    #   efficientnet-pytorch
    #   kornia
    #   pretrainedmodels
    #   pytorch-lightning
    #   timm
    #   torchmetrics
    #   torchvision
torchmetrics==0.9.3
    # via pytorch-lightning
torchsummary==1.5.1
    # via -r requirements/dev.in
torchvision==0.12.0+cu113
    # via
    #   -r requirements/dev.in
    #   pretrainedmodels
    #   segmentation-models-pytorch
    #   timm
tornado==6.2
    # via -r requirements/dev.in
tqdm==4.64.0
    # via
    #   -r requirements/dev.in
    #   pretrainedmodels
    #   pytorch-lightning
    #   segmentation-models-pytorch
typing-extensions==4.3.0
    # via
    #   black
    #   onnx
    #   pytorch-lightning
    #   qudida
    #   torch
    #   torchvision
urllib3==1.26.11
    # via
    #   requests
    #   sentry-sdk
wandb==0.12.21
    # via -r requirements/dev.in
werkzeug==2.2.1
    # via tensorboard
wheel==0.37.1
    # via tensorboard
yarl==1.7.2
    # via aiohttp
zipp==3.8.1
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
